Use new analytics backend, and fix releases (#66)
The automated releases have been broken since Google shutdown Universal
Analytics because the underlying data for the service was sourced
through Google Analytics. I've been working on a new custom analytics
backend which should allow for a lot more flexibility moving forward.
There will be more updates on that front soon, but this PR is the
minimal switch to get releases working again with fresh data.