Switch to TypeScript, tsup for builds, and hybrid ESM/CJS package
This is a fairly substantial update that lays the framework for the upcoming v2
release. The v1 code has been moved to a `v1-release` branch and will continue
to do nightly releases for the time being. All new development will be
targeting the `main` branch which will now correspond to v2.

The key changes here are:

- All development dependencies have been updated to current versions.
- The codebase has been ported to TypeScript and the package includes native
types.
- The package is now a hybrid module including both CommonJS and ECMAScript
module builds. These share a data file, so the impact on package size should
be minimal.
- Webpack has been replaced with `tsup` for builds. This made hybrid builds and
TypeScript bundle type declarations easier with less configuration.
- Unnecessary files have been removed from the package to reduce its size.
- The linting config was cleaned up and configured for TypeScript, and now runs
on CI.

Merges #68